+++ src/serialize19.lib/serialize19/serialize.std_string.h
@@ -6,7 +6,7 @@
 namespace serialize19 {
 
 template<Archive A, class T, class... Ts> void serialize(A& a, std::basic_string<T, Ts...>& string) {
-    auto size = string.size();
+    auto size = static_cast<uint32_t>(string.size());
     serialize(a, size);
     if (a.mode == ArchiveMode::Read) {
         string.resize(size);
